backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.0
orjson==3.9.10
pydantic-settings==2.9.1